+++ src/extension.ts
@@ -1,5 +1,7 @@
 "use strict";
 
+import net = require("net");
+
 import {
     workspace, ExtensionContext, Disposable, WorkspaceConfiguration
 } from "vscode";
@@ -7,8 +9,9 @@ import {
 import {
     LanguageClient, LanguageClientOptions,
     NodeModule,
+    StreamInfo,
     TransportKind
-} from "vscode-languageclient";
+} from "vscode-languageclient/node";
 
 import {
     debug, DebugSession, DebugConfigurationProvider, DebugAdapterDescriptorFactory,
@@ -16,6 +19,8 @@ import {
 } from "vscode";
 
 import getPort = require('get-port');
+
+let client: LanguageClient;
 
 export async function activate(context: ExtensionContext): Promise<void> {
 
@@ -44,11 +49,17 @@ export async function activate(context: ExtensionContext): Promise<void> {
         }
     };
 
-    let debuggerPort: number = await getPort();
+    const debuggerPort: number = await getPort();
+    const mitigatedDiagnostics: [string] = config.get<[string]>("Diagnostics.mitigated");
+    const suppressedDiagnostics: [string] = config.get<[string]>("Diagnostics.suppressed");
     let clientOptions: LanguageClientOptions = {
         documentSelector: ["wolfram"],
         initializationOptions: {
-            debuggerPort: debuggerPort
+            debuggerPort: debuggerPort,
+            diagnosticsOverrides: {
+                mitigated: mitigatedDiagnostics,
+                suppressed: suppressedDiagnostics
+            }
         }
     };
 
@@ -56,13 +67,22 @@ export async function activate(context: ExtensionContext): Promise<void> {
     // client can be deactivated on extension deactivation
 
     // create the language client and start the client.
-    context.subscriptions.push(new LanguageClient("WolframLanguageServer",
-        "Wolfram Language Server", serverOptions, clientOptions).start());
+    client = new LanguageClient("WolframLanguageServer",
+        "Wolfram Language Server", serverOptions, clientOptions);
     // register debug type "dap-wl"
     context.subscriptions.push(debug.registerDebugConfigurationProvider("dap-wl",
         new WolframDebugConfigProvider()));
     context.subscriptions.push(debug.registerDebugAdapterDescriptorFactory("dap-wl",
         new WolframDebugAdapterDescriptorFactory(debuggerPort)));
+
+    client.start();
+}
+
+export function deactivate(): Thenable<void> | undefined {
+    if (!client) {
+        return undefined;
+    }
+    return client.stop();
 }
 
 class WolframDebugConfigProvider implements DebugConfigurationProvider { }